requirements.txt
streamlit==1.39.0
gTTS==2.5.3
requests>=2.32.3
Pillow>=10.4.0
imageio-ffmpeg>=0.4.9
